refactor(updater): migrate localapiclient component to v2
Move from V1 structure (component.go + localapiclientimpl/) to V2
structure (def/component.go, impl/, fx/fx.go) and update callers.
