+++ comp/updater/localapiclient/def/component.go
